# api-docs.yaml
openapi: 3.1.0
info:
  title: WMU Developer Club's Financial API
  description: An API to interact with the backend of WMU Developer Club's database
    to manage finances and budgeting.
  contact:
    name: API Support Technician
    email: bluewheatthin@gmail.com
  version: v1.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /requests/search:
    put:
      tags:
      - Requests
      summary: Filters through the requests based on specified values
      description: "Takes in a JSON array, where each element is a Filter object consisting\
        \ of the column to filter by, the operation to filter based on, and the desired\
        \ value, and returns all of the rows in the Requests table which match the\
        \ Filter objects"
      operationId: filterRequests
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Filter"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /requests/edit/id={id}:
    put:
      tags:
      - Requests
      summary: Edits a request in the Requests table
      description: "Takes in a JSON object and the id of the request to edit, and\
        \ edits that Request in the Requests table with the new values provided. Returns\
        \ the object on success"
      operationId: editRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Request"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /requests/delete/id={id}:
    put:
      tags:
      - Requests
      summary: Deletes an request from the Requests table
      description: Modifies the deleted column of the request based on the id provided
        to be 1
      operationId: deleteRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /requests/approve/id={id}_val={val}:
    put:
      tags:
      - Requests
      summary: Toggles the approved flag for an request
      description: "Using the id provided, it will toggle the approved flag for an\
        \ request to either 1 or 0"
      operationId: approveRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: val
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /requests/create:
    post:
      tags:
      - Requests
      summary: Adds a request to the Requests table
      description: Takes in a JSON object and adds that Request to the Requests table.
        Returns the object on success
      operationId: createRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Request"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    Filter:
      type: object
      properties:
        col:
          type: string
          description: The column to filter by
          example: feeFlag
        op:
          type: string
          description: The operator to filter by
          example: leq
        val:
          type: object
          description: The value to filter by
          example: 5
      required:
      - col
      - op
      - val
    Request:
      type: object
      properties:
        id:
          type: integer
          format: int32
        communityName:
          type: integer
          format: int32
          description: The community name for the requestee
          example: Web team
        requesteeUser:
          type: string
          description: The person who is submitting the request
          example: Khang
        itemName:
          type: string
          description: The name of the item being requested
          example: Napkins
        approval:
          type: integer
          format: int32
          description: "Whether or not the request has been approved, denied, or neither"
          example: 1
        quantity:
          type: integer
          format: int32
          description: The number of items that is being requested
          example: 12
        pricePerUnit:
          type: number
          description: The price for one unit of the item being requested
          example: 14.99
        deadline:
          type: string
          format: date-time
          description: When the item is needed by
          example: 2026-05-23T05:00:00
        purpose:
          type: string
          description: Why the requestee is asking for the item
          example: the attendees of event x need food
        deleted:
          type: integer
          format: int32
          description: A flag for if the request has been deleted or not
          example: 0
      required:
      - communityName
      - deadline
      - deleted
      - itemName
      - pricePerUnit
      - purpose
      - quantity
      - requesteeUser
...